+++ pkg/apis/externaldns/types.go
@@ -14,10 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package config
+package externaldns
 
 import (
-	"fmt"
 	"os"
 
 	"github.com/spf13/pflag"
@@ -34,7 +33,7 @@ type Config struct {
 	LogFormat  string
 }
 
-// NewConfig returns new Configuration object
+// NewConfig returns new Config object
 func NewConfig() *Config {
 	return &Config{}
 }
@@ -46,12 +45,4 @@ func (cfg *Config) ParseFlags() {
 	flags.StringVar(&cfg.LogFormat, "log-format", "text", "log format output. options: [\"text\", \"json\"]")
 	flags.BoolVar(&cfg.Debug, "debug", false, "debug mode")
 	flags.Parse(os.Args)
-}
-
-// Validate custom validation for flags aside from flag library provided
-func (cfg *Config) Validate() error {
-	if cfg.LogFormat != "text" && cfg.LogFormat != "json" {
-		return fmt.Errorf("unsupported log format: %s", cfg.LogFormat)
-	}
-	return nil
 }
