+++ pkg/apis/externaldns/validation/validation.go
@@ -1,0 +1,31 @@
+/*
+Copyright 2017 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package validation
+
+import (
+	"fmt"
+	"github.com/kubernetes-incubator/external-dns/pkg/apis/externaldns"
+)
+
+// ValidateConfig performs validation on the Config object
+func ValidateConfig(cfg *externaldns.Config) error {
+	// TODO: Should probably return field.ErrorList
+	if cfg.LogFormat != "text" && cfg.LogFormat != "json" {
+		return fmt.Errorf("unsupported log format: %s", cfg.LogFormat)
+	}
+	return nil
+}
