+++ pkg/apis/externaldns/validation/validation_test.go
@@ -14,25 +14,26 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package config
+package validation
 
 import (
+	"github.com/kubernetes-incubator/external-dns/pkg/apis/externaldns"
 	"testing"
 )
 
 func TestValidateFlags(t *testing.T) {
-	cfg := NewConfig()
+	cfg := externaldns.NewConfig()
 	cfg.LogFormat = "test"
-	if err := cfg.Validate(); err == nil {
+	if err := ValidateConfig(cfg); err == nil {
 		t.Errorf("unsupported log format should fail: %s", cfg.LogFormat)
 	}
 	cfg.LogFormat = ""
-	if err := cfg.Validate(); err == nil {
+	if err := ValidateConfig(cfg); err == nil {
 		t.Errorf("unsupported log format should fail: %s", cfg.LogFormat)
 	}
 	for _, format := range []string{"text", "json"} {
 		cfg.LogFormat = format
-		if err := cfg.Validate(); err != nil {
+		if err := ValidateConfig(cfg); err != nil {
 			t.Errorf("supported log format: %s should not fail", format)
 		}
 	}
